2_backend/requirements.txt
flask 
flask-cors 
pyserial 
python-dotenv 
